(DiamondLightSource/hyperion#1077) Move detector into separate package as per review comment
