+++ src/dodal/devices/detector/__init__.py
@@ -1,0 +1,2 @@
+# export GDA detector names for convenience
+from dodal.devices.detector.detector import *  # noqa: F403
