+++ src/dodal/devices/detector/det_dim_constants.py
