+++ src/dodal/devices/detector/det_dist_to_beam_converter.py
