+++ tests/devices/unit_tests/detector/test_detector.py
