+++ tests/devices/unit_tests/test_beam_converter.py
@@ -1,7 +1,7 @@
 import pytest
 from mockito import when
 
-from dodal.devices.det_dist_to_beam_converter import (
+from dodal.devices.detector.det_dist_to_beam_converter import (
     Axis,
     DetectorDistanceToBeamXYConverter,
 )
